heif: Move multi-XMP test to testdata fixture
